+++ click_extended/transform/__init__.py
@@ -1,11 +1,35 @@
 """Initialization file for the `click_extended.transform` module."""
 
 from click_extended.transform.as_path import as_path
-from click_extended.transform.to_lowercase import to_lowercase
-from click_extended.transform.to_uppercase import to_uppercase
+from click_extended.transform.to_case import (
+    to_camel_case,
+    to_dot_case,
+    to_flat_case,
+    to_kebab_case,
+    to_lower_case,
+    to_meme_case,
+    to_pascal_case,
+    to_path_case,
+    to_screaming_snake_case,
+    to_snake_case,
+    to_title_case,
+    to_train_case,
+    to_upper_case,
+)
 
 __all__ = [
     "as_path",
-    "to_lowercase",
-    "to_uppercase",
+    "to_camel_case",
+    "to_dot_case",
+    "to_flat_case",
+    "to_kebab_case",
+    "to_lower_case",
+    "to_meme_case",
+    "to_pascal_case",
+    "to_path_case",
+    "to_screaming_snake_case",
+    "to_snake_case",
+    "to_title_case",
+    "to_train_case",
+    "to_upper_case",
 ]